rag_chatbot_k8_test/sync_docs/requirements.txt
langchain-openai
langchain_community
langchain_core
fastapi
uvicorn
dotenv
scikit-learn
requests
